+++ package.json
@@ -151,9 +151,9 @@
   "scripts": {
     "vscode:prepublish": "pnpm run js-yaml && pnpm run package",
     "compile": "pnpm run compile:peggy && pnpm run check-types && pnpm run lint && node esbuild.js",
-    "compile:peggy": "peggy -c ./src/pegconfig.json",
+    "compile:peggy": "peggy --dts --cache src/parser.peggy",
     "watch": "npm-run-all -p watch:*",
-    "watch:peggy": "peggy -w -c ./src/pegconfig.json",
+    "watch:peggy": "peggy -w --dts --cache src/parser.peggy",
     "watch:esbuild": "node esbuild.js --watch",
     "watch:tsc": "tsc --noEmit --watch --project tsconfig.json",
     "package": "pnpm run compile:peggy && pnpm run check-types && pnpm run lint && node esbuild.js --production",
@@ -169,7 +169,7 @@
     "js-yaml": "npm-run-all js-yaml:*",
     "js-yaml:snippets": "js-yaml syntaxes/igorpro.snippets.yaml > syntaxes/igorpro.snippets.json",
     "js-yaml:tmLanguage": "js-yaml syntaxes/igorpro.tmLanguage.yaml > syntaxes/igorpro.tmLanguage.json",
-    "js-yaml:apiReference": "js-yaml syntaxes/igorpro.apiReference.yaml > syntaxes/igorpro.apiReference.json"
+    "js-yaml:builtIns": "js-yaml syntaxes/igorpro.builtIns.yaml > syntaxes/igorpro.builtIns.json"
   },
   "dependencies": {
     "estraverse": "^5.3.0",
@@ -184,9 +184,9 @@
     "@types/vscode": "^1.100.0",
     "@typescript-eslint/eslint-plugin": "^8.31.1",
     "@typescript-eslint/parser": "^8.31.1",
-    "@vscode/test-cli": "^0.0.10",
+    "@vscode/test-cli": "^0.0.11",
     "@vscode/test-electron": "^2.5.2",
-    "@vscode/test-web": "^0.0.69",
+    "@vscode/test-web": "^0.0.71",
     "assert": "^2.1.0",
     "esbuild": "^0.25.3",
     "eslint": "^9.25.1",
@@ -196,6 +196,5 @@
     "npm-run-all": "^4.1.5",
     "peggy": "^5.0.3",
     "typescript": "^5.8.3"
-  },
-  "packageManager": "pnpm@10.4.1+sha512.c753b6c3ad7afa13af388fa6d808035a008e30ea9993f58c6663e2bc5ff21679aa834db094987129aa4d488b86df57f7b634981b2f827cdcacc698cc0cfb88af"
+  }
 }
